+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - an ArkScript program is no longer a single bloc, but can be composed of multiple bloc, thus we don't need to use a single big {} or (begin) bloc for all the program
 - enhancing lexer and parser error messages
 - else clause in if constructions is now optional
+- updating error messages in the VM
 
 ### Removed
 - removed `fmt/format` from our dependencies